+++ Tests/Integration/MSFT_ScheduledTask.Integration.Tests.ps1
@@ -36,6 +36,7 @@ try
             AtLogon           = 'ScheduledTaskLogon'
             AtStartup         = 'ScheduledTaskStartup'
             ExecuteAs         = 'ScheduledTaskExecuteAs'
+            OnEvent           = 'ScheduledTaskOnEvent'
         }
 
         $configData = @{