+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Changed the scope from Global to Script ComputerManagementDsc.Common.Tests.ps1
 - ScheduledTask:
   - Added support for event based triggers, implemented using the ScheduleType OnEvent
+    fixes [Issue #167](https://github.com/PowerShell/ComputerManagementDsc/issues/167)
 
 ## 5.1.0.0