+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   ComputerManagementDsc.Common module to match what TimeZone requires.
   It was not exported in ComputerManagementDsc.ResourceHelper and not
   used.
+- Add `server` parameter to `Computer` resource - fixes [Issue #161](https://github.com/PowerShell/ComputerManagementDsc/issues/161)
 
 ## 5.0.0.0